Settings window, thickness

+++ SortAnalyzer/include/canvas.h
@@ -16,8 +16,8 @@ public:
 
     void setImage (Image* image);
     void setPixel (size_t x, size_t y, const Pixel& pixel);
-    void drawSquare (double center_x, double center_y, int a, const Pixel& pixel);
-    void drawLine (Point2d p1, Point2d p2, const Pixel& pixel, int thickness);
+    void drawSquare (double center_x, double center_y, size_t a, const Pixel& pixel);
+    void drawLine (Point2d p1, Point2d p2, const Pixel& pixel, size_t thickness);
 
     double getRatio (bool is_vertical) override;
     void scrollOnceUp    (bool is_vertical) override;
